+++ transport/grpc/handler/types.go
@@ -2,7 +2,7 @@ package grpcHandler
 
 import (
 	"github.com/alhamsya/boilerplate-go/lib/helpers/config"
-	"github.com/alhamsya/boilerplate-go/transport/inter/grpc/routers"
+	"github.com/alhamsya/boilerplate-go/transport/grpc/routers"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/health/grpc_health_v1"
 )
